refactor: 게임 진행 상태 추가
